Establishing Connection If isConnected is false, indicating that there is no active connection to MongoDB, the function attempts to connect to MongoDB using mongoose.connect(process.env.MONGO_URI!). Here:
process.env.MONGO_URI refers to an environment variable that should contain the URI string necessary to connect to the MongoDB instance. The ! asserts that this environment variable will not be null or undefined.
mongoose.connect() is an asynchronous operation that returns a promise which resolves to a mongoose.Connection object upon successful connection.