Interfaces Two interfaces are defined:
ChatInterface:

Defines the structure of the Chat document.
It has two properties:
participants: An array of Types.ObjectId which represents references to User documents.
messages: An array of Types.ObjectId which represents references to Message documents.
ChatDocument:

Extends ChatInterface and Document.
Adds two additional properties:
createdAt: A Date type indicating when the document was created.
updatedAt: A Date type indicating when the document was last updated.