Schema Definition A Mongoose schema (chatModel) is defined using TypeScript's generic support:
The schema defines two fields:
participants: An array of objects where each object is expected to be a mongoose.Schema.Types.ObjectId, referencing documents in the User collection (ref: 'User').
messages: An array of objects where each object is expected to be a mongoose.Schema.Types.ObjectId, referencing documents in the Message collection (ref: 'Message').
{ timestamps: true }: This option tells Mongoose to automatically manage createdAt and updatedAt fields for each document.