Interfaces Two interfaces are defined:
MessageInterface:

Defines the structure of the Message document.
It has five properties:
senderId: Either a Types.ObjectId or a PopulatedDoc<UserDocument>. This indicates that senderId can be either an ObjectId referencing a document in the User collection or an already populated UserDocument.
receiverId: Similar to senderId, indicating the recipient of the message.
content: A string representing the content of the message. It is required (required: true).
messageType: A string literal union ('text' | 'image') specifying the type of message.
opened: A boolean indicating whether the message has been opened.
MessageDocument:

Extends MessageInterface and Document.
Adds two additional properties:
createdAt: A Date type indicating when the document was created.
updatedAt: A Date type indicating when the document was last updated.