Schema Definition A Mongoose schema (messageModel) is defined using TypeScript's generic support:
The schema defines five fields:

senderId: An ObjectId referencing documents in the User collection (ref: 'User').
receiverId: Also an ObjectId referencing documents in the User collection (ref: 'User').
content: A String type field that is required (required: true).
messageType: A String type field that is required and constrained to be either 'text' or 'image' (enum: ['text', 'image']).
opened: A Boolean type field that defaults to false.