Interfaces An interface UserInterface is defined:
Defines the structure of the User document.
It has four properties:
username: A string representing the username of the user. It is required (required: true) and must be unique (unique: true).
fullname: A string representing the full name of the user. It is required (required: true).
email: A string representing the email address of the user. It is required (required: true) and must be unique (unique: true).
profilePhoto: A string representing the URL or path to the user's profile photo. It has a default value of an empty string ("").
Interface Extension
An interface UserDocument is defined, extending UserInterface and Document:

UserDocument incorporates all properties from UserInterface.
Adds two additional properties:
createdAt: A Date type indicating when the document was created. This is provided by Mongoose's timestamps option.
updatedAt: A Date type indicating when the document was last updated. Also provided by Mongoose's timestamps option.