deleteChatMessages Function Function Definition: deleteChatMessages is an asynchronous function that accepts userId as a parameter, representing the ID of the user whose chat messages are to be deleted.
NoStore: Similar to sendSnapMessage, noStore() likely ensures that caching is bypassed for this operation.

Database Connection: connectDB() is called to establish a connection to the MongoDB database.

Authentication: auth() is called to authenticate the user. If successful, authUser.user is retrieved.

Chat Retrieval: The function attempts to find a Chat document where both user._id and userId are participants.

Message Deletion: If a Chat document is found, it retrieves the IDs of all messages in the chat (messageIdsInString) and deletes them using Message.deleteMany().

Chat Deletion: After deleting all messages, the entire Chat document is deleted using Chat.deleteOne().

Revalidation: revalidatePath(/chat/${userId}) triggers a revalidation of the cached data associated with the chat page of userId.

Redirect: After successful deletion, the function redirects the user to /chat.